+++ CHANGELOG.md
@@ -32,6 +32,7 @@ CHANGELOG
 
 ### Tests
 - Replace calls to deprecated/undocumented `logging.warn` with `logging.warning` (by Sebastian Wagner, fixes #2013).
+- `intelmq.tests.bots.experts.rdap.test_expert`: Declare cache use, fixes build failures (by Sebastian Wagner, fixes #2014).
 
 ### Tools